workers/agent_analysis/pyproject.toml
[project]
name = "agent-analysis-worker"
version = "0.1.0"
description = "AI report interpretation and intervention worker"
requires-python = ">=3.12"
dependencies = [
    "aio-pika>=9.4.0",
    "dashscope==1.25.2",
    "httpx>=0.28.1",
    "langchain==1.1.3",
    "langchain-community==0.4.1",
    "langchain-core==1.1.2",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.2",
]
